Move files into groth16 mod
